Prevent using same request without changing captcha
If someone enter proper captcha on one page, and have another page which just check was captcha valid, he can make unlimited number of requests with simply refreshing the page, without need to enter new captcha. This happens because you re-create hash only when image is created, but if someone don't show captcha, and check on the same page, then Hash will be created on first page, and on second it will be always valid.